+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.2.1
+- [#9](https://github.com/felixgilioli/felix-utils/issues/9) - added StringUtils.isNotBlank.
+- [#8](https://github.com/felixgilioli/felix-utils/issues/8) - added StringUtils.isBlank.
+
 ## 1.1.1
 - [#2](https://github.com/felixgilioli/felix-utils/issues/2) - added ObjectUtils.checkArgument.
 - [#1](https://github.com/felixgilioli/felix-utils/issues/1) - Added method ListUtils.findFirst.